delete app for retreival apps
